Add zsh tag (#125)
For zsh to recognize completion file, the file in question needs to
include a tag `#compdef name`, see [1]:

````
When compinit is run, it searches all such files accessible via
fpath/FPATH and reads the first line of each of them. This line
should contain one of the tags described below. Files whose first line
does not start with one of these tags are not considered to be part of
the completion system and will not be treated specially.
````

[1] http://zsh.sourceforge.net/Doc/Release/Completion-System.html#Autoloaded-files