Add support for multiple user data files
The option --user-data-from-file can now be used multiple times on a
"server create" command. If this option is specified more than once, all
files will be opened and put into a multipart MIME message which will
then be sent to the API. The cloud-init can understand this MIME
message and unpack it.