allow specifying default columns for each sub command
I am a heavy user of the hcloud-cli. However, I usually change the columns that are displayed and don't like the default and would prefer specifying the default columns used in the config file.
Note: This PR adds on top of the work of #421 and should not be merged
before #421.

Currently, I always type `hcloud server list -o 'columns=name,status,ipv6,private_net'` to only get the columns that I'm interested in

since the `-o` flag is part of the `list` subcommand and not the `hcloud` base command (like in many other software that makes use of the Cobra framework),
I cannot simply `alias hcloud-"hcloud -o 'columns='"` in my shell to default the columns.

This commit adds a `[defaults]` section to the config toml which allows to override the default sorting behavior for individual `hcloud` subcommands.

The following behavior is used to determine the displayed columns

| config file | flag    | result                      |
|-------------|---------|-----------------------------|
| not-set     | not-set | display default columns     |
| not-set     | set     | display -o flag columns     |
| set         | not-set | display config file columns |
| set         | set     | display -o flag columns     |