feat(completion): read network zones from API
We previously hard-coded the available network zones for autocompletion
purposes. This makes it easy to forget to add them, or only update the list
in one location (as you can see, it's currently missing us-east for the
loadbalancer create call).

In this PR we replace the hardcoded list with an api call that retrieves
all available network zones from the API. As there is no dedicated
`GET /v1/network_zones` endpoint, we use the network zone referenced in
the `GET /v1/locations` response.