fix(ci): race-condition in signing macos binaries
By default the keychain that is used to sign macOS binaries is only
unlocked for 300s. This caused the release process to fail, as building
all binaries took longer than the 300s, and subsequent signing failed.

The new timeout is 3600s (1h), which should be enough to build anything we
want. At the moment building all binaries takes ~8m, so we have a buffer of
52m or 650%.