fix(ci): race-condition in signing macos binaries (#433)
* fix(ci): race-condition in signing macos binaries

By default the keychain that is used to sign macOS binaries is only
unlocked for 300s. This caused the release process to fail, as building
all binaries took longer than the 300s, and subsequent signing failed.

The new timeout is 3600s (1h), which should be enough to build anything we
want. At the moment building all binaries takes ~8m, so we have a buffer of
52m or 650%.

* fix(ci): show logs of gon hooks

This makes debugging any issues encountered in these steps easier.

* fix(ci): use similar goreleaser settings to tagged release builds

In release builds we use `--parallelism=2` and the latest goreleaser version.

We should use the same settings in our dev builds.