fix: send debug output to stderr
When enabling debug output (`HCLOUD_DEBUG=true`), this was being
printed to stdout by default. If one was using the output of the
command in a script, this would break that usage.

By outputting to stderr, we can make sure that the "normal" output
still works as expected and can be used in subsequent steps.

For example:

    export HCLOUD_DEBUG=true
    IP=$(hcloud server ip hccm-dev-1)
    ping $IP