ci: fix release assets not available
The default token GH_TOKEN in workflows does not trigger new events.
We rely on the "tag" event created by release-please to build our
release assets.

By using a token for "hcloud-bot", the event will be properly created.

We explicitly only run the workflow job if the repository is the main
repo, as otherwise this workflow will fail in forks because the secret
is not available.