fix: static resource list commands only return first 50 entries
The list commands for datacenters, load balancer types, locations and server types used List() instead of AllWithOpts() to fetch resources. List() only returns a single page of resources at a time (50 entries in this case), while AllWithOpts() returns all of them. This does not make a difference at the moment, but could be a problem in the future.

Similar to #574