ci: replace gon with rcodesign, switch to ubuntu runner (#738)
This PR replaces the unmaintained `mitchellh/gon` with
[`rcodesign`](https://gregoryszorc.com/docs/apple-codesign/stable),
which makes the CI pipeline able to run on Linux. This should make the
build and release process faster.

The Apple Developer ID and password are not used anymore (this was not
clear before). Also, the Apple code signing certificate has been moved
to a secret.


Closes #676

---------

Co-authored-by: Jonas L <jooola@users.noreply.github.com>