docs(server): update ssh command usage string (#787)
The current usage string creates confusion about whether it is possible
to pass flags to the ssh binary or not (See #782). The new usage strings
explicitly states that this is possible. It also states that it is
possible to use a double dash (`--`) to separate hcloud flags from ssh
flags.