feat: better error messages on malformed option value (#798)
Previously, if an option was configured with the wrong type (e.g.
`string` was configured with an `int` value) or other parsing errors
like malformed duration/boolean values occured, `Option[T].Get()` would
panic. It now returns an error that can properly be displayed to the
user and falls back to the default value.