fix(config): ordering of list option values not preserved (#805)
Originally, options with a slice value were treated as sets instead of
lists. This was because options like `default-ssh-keys` do not care
about the order of keys. However, the order for example is important for
the default sort column options, so it needs to be preserved. This PR
fixes this.