feat: new APIs for traffic pricings  (#827)
The API has been updated to provide a better insight and more
flexibility for displaying the pricing of traffic for servers and load
balancers.

In addition to the new fields, the old fields are deprecated and will be
set to `null` in the API on 2024-08-05.

In the CLI this mostly affects the `hcloud server-type list` command,
where we have previously shown a column `traffic`. This is no longer
possible, because there is no single included traffic for a server type,
and instead this depends on the location. We have set the field to `-`
and removed it from the default columns. The information is still
available in `hcloud server-type describe <server-type-name>`, now in
the "Pricing" section of the output.

This also affects the JSON and YAML output of both `hcloud server-type
list` and `hcloud server-type describe` as we use the API schema for
this.

You can learn more about this change in [our
changelog](https://docs.hetzner.cloud/changelog#2024-07-25-cloud-api-returns-traffic-information-in-different-format).