ci: add coverpkg flag for testing (#815)
This PR adds the `-coverpkg=./...` flag so that coverage is not only
restricted to lines in the same package as the running test. This flag
is needed for end to end testing. Adding the flag now makes it easier
for us to compare the coverage of e2e tests with the main branch.