ci: build and test with Go 1.23 (#871)
This PR makes the CI pipeline use Go 1.23 for building and testing. Only
setting the `toolchain` in `go.mod` is not enough, because the [setup-go
action does not support the `toolchain` directive
yet](https://github.com/actions/setup-go/issues/457), we have to set the
Go version to 1.23 in all setup-go CI steps manually.
We could also set the `go` version in `go.mod` to 1.23 since we do not
have any dependents, but that would be misusing the `go` directive,
since it is only intended to specify the minimum supported language
version of the project.