refactor: tablewriter that can handle unicode (#995)
The previously used `text/tabwriter` does not even attempt to handle
Unicode characters that may have a default width != 0. This causes
issues in tables where any column has such a Unicode character (like
emojis), as the alignment for that row would be broken.

We already use the new package for generating our docs, so this "only"
moves it from tools to production dependency.

I tried to copy the previous style as much as possible and left the
surrounding code as is. There is some potential for simplification
because the new package can handle a lot of the column/view logic for
us.