chore: fix config option default (#1019)
This PR removes the default value for the `Config` option and moves the
logic into the config parsing itself.
This is so that the user's home path is not shown in help output. This
helps with reproducibility and also allows generating documentation.

In preparation for https://github.com/hetznercloud/cli/pull/1018