fix: config option flags sometimes not parsed correctly (#1025)
See https://github.com/hetznercloud/cli/issues/845 for a description of
the error

While parsing the config's `FlagSet`, if the parser reached a flag that
was not a config option, it would terminate early and return an unknown
flag error. Since the error was ignored silently, this behavior was not
apparent to the user.

This PR fixes this behavior by using the `ParseErrorsWhitelist` to allow
unknown flags in the config's `FlagSet`. This also means we can now
return any other parsing errors properly instead of ignoring them.

Fixes #845