ci: only count coverage for internal package (#1076)
Currently we count code coverage when testing (using the `-coverpkg`)
across the entire repository and all of its subdirectories. This caused
problems in one of our CI jobs, where a Go installation was placed
inside of the repository in a hidden folder, causing `-coverpkg` to also
cover the entire Go source.
Since all of the code that needs testing is located in the `internal`
package, we should restrict our tests to only cover that package.